Clean-up some of the logic in DiscoverTargets and WatchedFileChange
