+++ Sources/SourceKitBazelBSP/RequestHandlers/BuildTargets/BazelTargetDiscoverer.swift
@@ -1,0 +1,66 @@
+import Foundation
+
+// MARK: - Logger
+
+private let logger = makeFileLevelBSPLogger()
+
+// MARK: - Public API
+
+public enum TopLevelTarget: String, CaseIterable {
+    case iosApplication = "ios_application"
+    case iosUnitTest = "ios_unit_test"
+}
+
+public enum BazelTargetDiscovererError: LocalizedError, Equatable {
+    case noTargetsDiscovered
+
+    public var errorDescription: String? {
+        switch self {
+        case .noTargetsDiscovered: return "No targets discovered!"
+        }
+    }
+}
+
+public enum BazelTargetDiscoverer {
+    /// Discovers Bazel targets in a workspace matching the specified rule
+    /// types by running a Bazel query. To be used when the BSP is initialized
+    /// with no targets specified.
+    ///
+    /// - Parameters:
+    ///   - rules: The rule types to search for (e.g. ios_application, ios_unit_test)
+    ///   - bazelWrapper: The Bazel executable to use (defaults to "bazel")
+    /// - Returns: An array of discovered target labels
+    /// - Throws: DiscoverTargetsError.noTargetsDiscovered if no matching targets are found
+    public static func discoverTargets(
+        for rules: [TopLevelTarget] = TopLevelTarget.allCases,
+        bazelWrapper: String = "bazel",
+        commandRunner: CommandRunner? = nil
+    ) throws -> [String] {
+        let commandRunner = commandRunner ?? ShellCommandRunner()
+
+        logger.info("Discovering targets for rules: \(rules.map { $0.rawValue }.joined(separator: ", "))")
+
+        let query = rules.map {
+            "kind(\($0.rawValue), ...)"
+        }
+        .joined(separator: " + ")
+
+        let cmd = "\(bazelWrapper) query '\(query)' --output label"
+
+        let output = try commandRunner.run(cmd)
+
+        let discoveredTargets =
+            output
+            .trimmingCharacters(in: .whitespacesAndNewlines)
+            .components(separatedBy: .newlines)
+            .map { $0.trimmingCharacters(in: .whitespacesAndNewlines) }
+            .filter { !$0.isEmpty }
+
+        if discoveredTargets.isEmpty {
+            logger.error("No targets discovered!")
+            throw BazelTargetDiscovererError.noTargetsDiscovered
+        }
+
+        return discoveredTargets
+    }
+}
